+++ example/pages/swipe-cell/index.js
@@ -1,0 +1,21 @@
+import Page from '../../common/page';
+import Dialog from '../../dist/dialog/dialog';
+
+Page({
+  onClose(event) {
+    const { position, instance } = event.detail;
+    switch (position) {
+      case 'left':
+      case 'cell':
+        instance.close();
+        break;
+      case 'right':
+        Dialog.confirm({
+          message: '确定删除吗？'
+        }).then(() => {
+          instance.close();
+        });
+        break;
+    }
+  }
+});