+++ packages/panel/index.ts
@@ -1,13 +1,12 @@
 import { VantComponent } from '../common/component';
 
 VantComponent({
-  classes: ['footer-class'],
+  classes: ['header-class', 'footer-class'],
 
   props: {
     desc: String,
     title: String,
     status: String,
-    headerClass: String,
     useFooterSlot: Boolean
   }
 });