+++ packages/search/index.ts
@@ -7,8 +7,10 @@ VantComponent({
 
   props: {
     focus: Boolean,
+    error: Boolean,
     disabled: Boolean,
     readonly: Boolean,
+    inputAlign: String,
     showAction: Boolean,
     useActionSlot: Boolean,
     placeholder: String,